+++ packages/stream_video_flutter/lib/src/theme/components/components.dart
@@ -1,6 +1,7 @@
 export 'call_app_bar_theme.dart';
 export 'call_button_badge_theme.dart';
 export 'call_control_bar_theme.dart';
+export 'call_duration_badge_theme.dart';
 export 'call_participants_grid_theme.dart';
 export 'connection_quality_indicator_theme.dart';
 export 'desktop_screen_share_selector_theme.dart';
@@ -9,4 +10,3 @@ export 'lobby_view_theme.dart';
 export 'participant_label_theme.dart';
 export 'participant_tile_theme.dart';
 export 'picture_in_picture_theme.dart';
-export 'stream_call_duration_badge_theme.dart';